make the prodactivity tools useable by non-root
Issues:
Fixes #13

Problem:  For prototyping we were assuming root access in the docker worker
this assumption is not met in nightly regression runs so we needed to upgrade
to enable the publish_docker_test_container to function without root
privileges.

Analysis: In particular the docker socker '/var/run/docker.sock' belongs
to group 'staff' so the user (e.g. buildbot) needs to also belong to that
group.'

Tests:  Running tests from the "driver" repo in a nightly regression proves
these changes.   The changes should also be tested manually before acceptance.